backend/cloud-functions/party-3/requirements.txt
flask<3.0,>=1.0
firebase-admin
PyNaCl